Move `required` property to SelectMenu
