+++ src/main/java/net/dv8tion/jda/api/components/selections/EntitySelectMenu.java
@@ -38,7 +38,6 @@ import org.jetbrains.annotations.Unmodifiable;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.*;
 
 /**
@@ -134,14 +133,6 @@ public interface EntitySelectMenu extends SelectMenu, LabelChildComponent
     @Nonnull
     @Unmodifiable
     List<DefaultValue> getDefaultValues();
-
-    /**
-     * Whether the user must populate this select menu in Modals, or {@code null} if not set.
-     *
-     * @return Whether this menu must be populated, or null
-     */
-    @Nullable
-    Boolean isRequired();
 
     /**
      * Creates a new preconfigured {@link Builder} with the same settings used for this select menu.
@@ -477,22 +468,10 @@ public interface EntitySelectMenu extends SelectMenu, LabelChildComponent
         protected Component.Type componentType;
         protected EnumSet<ChannelType> channelTypes = EnumSet.noneOf(ChannelType.class);
         protected List<DefaultValue> defaultValues = new ArrayList<>();
-        protected Boolean required;
 
         protected Builder(@Nonnull String customId)
         {
             super(customId);
-        }
-
-        /**
-         * Whether the user must populate this select menu in Modals, or {@code null} if not set.
-         *
-         * @return Whether this menu must be populated, or null
-         */
-        @Nullable
-        public Boolean isRequired()
-        {
-            return required;
         }
 
         /**
@@ -659,24 +638,6 @@ public interface EntitySelectMenu extends SelectMenu, LabelChildComponent
 
             this.defaultValues.clear();
             this.defaultValues.addAll(values);
-            return this;
-        }
-
-        /**
-         * Configure whether the user must populate this select menu if inside a Modal.
-         * <br>This defaults to {@code true} in Modals when unset.
-         *
-         * <p>This only has an effect in Modals!
-         *
-         * @param required
-         *        Whether this menu is required
-         *
-         * @return The same builder instance for chaining
-         */
-        @Nonnull
-        public Builder setRequired(@Nullable Boolean required)
-        {
-            this.required = required;
             return this;
         }
