+++ src/main/java/net/dv8tion/jda/api/components/selections/StringSelectMenu.java
@@ -101,14 +101,6 @@ public interface StringSelectMenu extends SelectMenu, LabelChildComponent
     List<SelectOption> getOptions();
 
     /**
-     * Whether the user must populate this select menu in Modals, or {@code null} if not set.
-     *
-     * @return Whether this menu must be populated, or null
-     */
-    @Nullable
-    Boolean isRequired();
-
-    /**
      * Creates a new preconfigured {@link Builder} with the same settings used for this select menu.
      * <br>This can be useful to create an updated version of this menu without needing to rebuild it from scratch.
      *
@@ -153,7 +145,6 @@ public interface StringSelectMenu extends SelectMenu, LabelChildComponent
     class Builder extends SelectMenu.Builder<StringSelectMenu, StringSelectMenu.Builder>
     {
         private final List<SelectOption> options = new ArrayList<>();
-        private Boolean required = null;
 
         protected Builder(@Nonnull String customId)
         {
@@ -307,17 +298,6 @@ public interface StringSelectMenu extends SelectMenu, LabelChildComponent
         }
 
         /**
-         * Whether the user must populate this select menu in Modals, or {@code null} if not set.
-         *
-         * @return Whether this menu must be populated, or null
-         */
-        @Nullable
-        public Boolean isRequired()
-        {
-            return required;
-        }
-
-        /**
          * Configures which of the currently applied {@link #getOptions() options} should be selected by default.
          *
          * @param  values
@@ -393,24 +373,6 @@ public interface StringSelectMenu extends SelectMenu, LabelChildComponent
         {
             Checks.noneNull(values, "Values");
             return setDefaultOptions(Arrays.asList(values));
-        }
-
-        /**
-         * Configure whether the user must populate this select menu if inside a Modal.
-         * <br>This defaults to {@code true} in Modals when unset.
-         *
-         * <p>This only has an effect in Modals!
-         *
-         * @param required
-         *        Whether this menu is required
-         *
-         * @return The same builder instance for chaining
-         */
-        @Nonnull
-        public Builder setRequired(@Nullable Boolean required)
-        {
-            this.required = required;
-            return this;
         }
 
         /**
