+++ src/main/java/net/dv8tion/jda/internal/components/selections/StringSelectMenuImpl.java
@@ -31,20 +31,17 @@ import java.util.Objects;
 public class StringSelectMenuImpl extends SelectMenuImpl implements StringSelectMenu, LabelChildComponentUnion
 {
     private final List<SelectOption> options;
-    private final Boolean required;
 
     public StringSelectMenuImpl(DataObject data)
     {
         super(data);
         this.options = parseOptions(data.getArray("options"));
-        this.required = (Boolean) data.opt("required").orElse(null);
     }
 
     public StringSelectMenuImpl(String id, int uniqueId, String placeholder, int minValues, int maxValues, boolean disabled, List<SelectOption> options, Boolean required)
     {
-        super(id, uniqueId, placeholder, minValues, maxValues, disabled);
+        super(id, uniqueId, placeholder, minValues, maxValues, disabled, required);
         this.options = options;
-        this.required = required;
     }
 
     private static List<SelectOption> parseOptions(DataArray array)
@@ -77,24 +74,13 @@ public class StringSelectMenuImpl extends SelectMenuImpl implements StringSelect
         return Collections.unmodifiableList(options);
     }
 
-    @Override
-    public Boolean isRequired()
-    {
-        return required;
-    }
-
     @Nonnull
     @Override
     public DataObject toData()
     {
-        DataObject obj = super.toData()
+        return super.toData()
                 .put("type", Type.STRING_SELECT.getKey())
                 .put("options", DataArray.fromCollection(options));
-
-        if (required != null)
-            obj.put("required", required);
-
-        return obj;
     }
 
     @Override
